DEV: Remove deprecation and upgrade functionality
The `about.json` change made back in 2021 was invalid, which means anyone using this component is stuck on an old version and unable to update.

Having reviewed some use-cases, it seems there is still some value to this theme component, even with the new core site settings.

Therefore, this commit removes the (failed) deprecation, and updates the implementation to work with the new header.

Requires https://github.com/discourse/discourse/pull/28189